인증글 리스트, 등록 완료

+++ src/controller/admin/challenge/AdminChallengeDeleteController.java
@@ -23,6 +23,11 @@ public class AdminChallengeDeleteController extends HttpServlet {
 	
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		//로그인이 되어있지 않으면 리다이렉트
+		if(req.getSession().getAttribute("login")==null) {
+			resp.sendRedirect("/");
+			return;
+		}
 		adminChallengeService.delete(req);
 		resp.sendRedirect("/admin/challenge/list");
 	}
